Add 120Hz and 144Hz refresh rates
This patch adds 120 and 144HZ refresh rates because
Proton fullscreen patches only support 60Hz for fullscreen
The same applies to the wine virtual desktop which only
supports 60Hz and causes many games to have bad micro
stutters if the framerate exceeds the artifical 60Hz
while running with e.g. 144Hz natively.
With this patch freesync and 144Hz runs butter smooth
for displays that support it.