mingw: mingw-w64-headers 6.0.0 folder in 7.0.0 sources
The 7.0.0 sources contain a folder named 6.0.0 so this adds a patch to work
around that issue until either mingw fixes this issue on their end or the AUR
PKGBUILD adopts this or another workaround like renaming the folder.

mingw: mingw-w64-crt 6.0.0 folder in 7.0.0 sources

The 7.0.0 sources contain a folder named 6.0.0 so this adds a patch to work
around that issue until either mingw fixes this issue on their end or the AUR
PKGBUILD adopts this or another workaround like renaming the folder.

What's the enter key for again?

mingw: mingw-w64-crt patch sha256sum

Patch the checksum to the correct one

mingw: more 7.0.0 workarounds